进一步简化插件加载测试脚本
