+++ src/plugins/store/models.py
@@ -1,4 +1,19 @@
-from typing import TypedDict
+from typing import Any, TypedDict
+
+
+class PluginData(TypedDict):
+    """NoneBot 商店插件数据"""
+
+    module_name: str
+    project_link: str
+    name: str
+    desc: str
+    author: str
+    homepage: str
+    tags: list[Any]
+    is_official: bool
+    type: str
+    supported_adapters: list[str]
 
 
 class Metadata(TypedDict):
